feat(smartcar): SoC-Felder von INTEGER auf NUMERIC(5,2) migrieren
V91: smartcar_webhook_raw_log.soc_percent, V92: ev_trip.soc_start/soc_end.
Tesla liefert sub-percent SoC (z.B. 55.828%) - bisher durch Math.round()
auf Integer gerundet und Präzision verloren. Nötig für Phantom-Drain-Analyse.

+++ backend/src/main/java/com/evmonitor/application/InternalSmartcarWebhookLogRequest.java
@@ -10,7 +10,7 @@ public record InternalSmartcarWebhookLogRequest(
         Integer year,
         String triggersJson,
         String signalsJson,
-        Integer socPercent,
+        BigDecimal socPercent,
         BigDecimal odometerKm,
         String locationGeohash,
         BigDecimal outsideTempCelsius,
